Bye Bye java.awt

+++ src/main/java/me/shedaniel/rei/gui/widget/RecipeChoosePageWidget.java
@@ -7,6 +7,8 @@ package me.shedaniel.rei.gui.widget;
 
 import com.google.common.collect.Lists;
 import com.mojang.blaze3d.platform.GlStateManager;
+import me.shedaniel.math.api.Point;
+import me.shedaniel.math.api.Rectangle;
 import me.shedaniel.rei.RoughlyEnoughItemsCore;
 import me.shedaniel.rei.api.ConfigManager;
 import me.shedaniel.rei.gui.RecipeViewingScreen;
@@ -19,7 +21,6 @@ import net.minecraft.client.resource.language.I18n;
 import net.minecraft.client.util.Window;
 import net.minecraft.util.math.MathHelper;
 
-import java.awt.*;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
@@ -47,7 +48,7 @@ public class RecipeChoosePageWidget extends DraggableWidget {
     private static Point getPointFromConfig() {
         Window window = MinecraftClient.getInstance().window;
         ConfigObjectImpl.RelativePoint point = RoughlyEnoughItemsCore.getConfigManager().getConfig().getChoosePageDialogPoint();
-        return new Point((int) point.getX(window.getScaledWidth()), (int) point.getY(window.getScaledHeight()));
+        return new Point(point.getX(window.getScaledWidth()), point.getY(window.getScaledHeight()));
     }
     
     @Override
